requirements.txt
# Core dependencies
asyncpg==0.27.0
python-binance==1.0.17
pybit==5.6.2
python-dotenv==1.0.0
aiohttp==3.8.4

# Security
cryptography==41.0.7

# WebSocket
websockets==11.0.3

# Monitoring
prometheus-client==0.19.0

# Web API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Rate limiting
aiolimiter==1.1.0

# Utils
python-json-logger==2.0.7
tabulate==0.9.0